requirements.txt
jinja2
pyyaml
boto3